# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: InterviewAI's authenticated interview-practice API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerAccount
      tags: [auth]
      summary: Create a Supabase Auth account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Registration details are invalid
  /auth/login:
    post:
      operationId: loginAccount
      tags: [auth]
      summary: Start a session using email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Credentials are invalid
...
  /auth/me:
    get:
      operationId: getCurrentAccount
      tags: [auth]
      summary: Restore the current session
      responses:
        "200":
          description: Current authenticated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: No active session
...
  /roadmap/{id}:
    patch:
      operationId: updateRoadmapItem
      tags: [roadmap]
      summary: Update roadmap topic progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoadmapProgressInput"
      responses:
        "200":
          description: Updated roadmap topic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadmapItem"
components:
...
  schemas:
...
    AuthUser:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        fullName: { type: string }
        username: { type: string }
      required: [id, email, fullName, username]
    RegisterInput:
      type: object
      properties:
        fullName: { type: string, minLength: 2, maxLength: 100 }
        username: { type: string, minLength: 3, maxLength: 32 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
      required: [fullName, username, email, password]
    LoginInput:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1, maxLength: 128 }
      required: [email, password]
    AuthResult:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        message: { type: string }
        emailConfirmationRequired: { type: boolean }
      required: [user, message, emailConfirmationRequired]
...
    RoadmapItem:
      type: object
      properties:
        id: { type: string }
        topic: { type: string }
        description: { type: string }
        priority: { type: string }
        practiceTask: { type: string }
        estimatedMinutes: { type: integer }
        status: { type: string, enum: [Not Started, In Progress, Completed] }
        orderIndex: { type: integer }
      required: [id, topic, description, priority, practiceTask, estimatedMinutes, status, orderIndex]
    RoadmapProgressInput:
      type: object
      properties:
        status: { type: string, enum: [Not Started, In Progress, Completed] }
      required: [status]